% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalingUtils.R
\name{scalingType}
\alias{scalingType}
\title{Utility for handling user friendly scaling --- None exported}
\usage{
scalingType(
  scaling = c("none", "sites", "species", "symmetric"),
  correlation = FALSE,
  hill = FALSE
)
}
\arguments{
\item{scaling}{character or numeric; which type of scaling is required? Numeric values are returned unaltered}

\item{correlation}{logical; should correlation-like scores be returned?}

\item{hill}{logical; should Hill's scaling scores be returned?}
}
\description{
Convert user-friendly descriptions of scalings to numeric codes used by \code{scores} to date.
}
