% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_wrappers.R
\name{t_wrappers}
\alias{t_wrappers}
\alias{t_cca}
\alias{t_rda}
\alias{t_metaMDS}
\alias{t_vegdist}
\alias{t_decostand}
\alias{t_adonis2}
\alias{t_anosim}
\alias{t_mrpp}
\alias{t_specaccum}
\alias{t_rarecurve}
\alias{t_rarefy}
\alias{t_rrarefy}
\alias{t_drarefy}
\title{Tidy Wrappers for vegan Functions}
\usage{
t_cca(data, formula, ...)

t_rda(data, formula, ...)

t_metaMDS(data, ...)

t_vegdist(data, method = "bray", ...)

t_decostand(data, method, MARGIN, ...)

t_adonis2(data, formula, ...)

t_anosim(data, grouping, ...)

t_mrpp(data, grouping, ...)

t_specaccum(data, ...)

t_rarecurve(data, ...)

t_rarefy(data, sample, ...)

t_rrarefy(data, sample)

t_drarefy(data, sample)
}
\arguments{
\item{data}{A data frame, tibble, or matrix. For constrained ordination, this is
either the community data (if formula is missing) or the environmental data (if
formula is provided).}

\item{formula}{A formula specifying the model.}

\item{...}{Additional arguments passed to the underlying function.}

\item{method}{Dissimilarity index or transformation method.}

\item{MARGIN}{Margin on which the operation is applied (1 for rows, 2 for columns).}

\item{grouping}{A grouping vector or factor.}

\item{sample}{Size of the sample to draw.}
}
\description{
These functions are tidyverse-compatible wrappers for the core functions of
the \pkg{tidyvegan} package. They are designed to be pipe-friendly by taking
a data frame, tibble, or matrix as their first argument.
}
